requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
cryptography==42.0.7
Django==5.0.3
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jwt==1.3.1
mysqlclient==2.2.4
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
zhipuai==2.1.0
sqlalchemy==2.0.30